feat(morph): add Morph filter with image morphing and video export
- Introduce MorphDisplay component for morphing between two images with preview
- Add MorphFilter with controls for morph amount, frames, fps, and transition type
- Implement shaders placeholder for morph effect
- Support video generation and export using FFmpeg and GIF fallback
- Update UI to replace MorphApp with Morph filter
- Add gif.js typings and FFmpeg loader utility
- Update store and package dependencies accordingly
