+++ package.json
@@ -19,6 +19,7 @@
     "@ffmpeg/util": "^0.12.2",
     "@pixi/filter-displacement": "^7.4.3",
     "@types/node": "^24.0.15",
+    "gif.js": "^0.2.0",
     "pixi.js": "^8.11.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
